requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
motor==3.6.0
numpy==2.2.1
openai==1.59.6
opencv-python==4.10.0.84
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.9.2
PyMuPDF==1.25.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
